refactor: resolve requested changes
