+++ serverhost-runtime/src/main/kotlin/app/simplecloud/droplet/serverhost/runtime/environment/BuildContext.kt
@@ -1,0 +1,6 @@
+package app.simplecloud.droplet.serverhost.runtime.environment
+
+enum class BuildContext {
+    STARTUP,
+    TRIGGER;
+}
