+++ serverhost-runtime/src/main/kotlin/app/simplecloud/droplet/serverhost/runtime/environment/GroupRuntime.kt
@@ -1,4 +1,4 @@
-package app.simplecloud.droplet.serverhost.runtime.runner
+package app.simplecloud.droplet.serverhost.runtime.environment
 
 import org.spongepowered.configurate.objectmapping.ConfigSerializable
