+++ serverhost-runtime/src/main/kotlin/app/simplecloud/droplet/serverhost/runtime/environment/GroupRuntimeDirectory.kt
@@ -1,4 +1,4 @@
-package app.simplecloud.droplet.serverhost.runtime.runner
+package app.simplecloud.droplet.serverhost.runtime.environment
 
 import app.simplecloud.controller.shared.group.Group
 import app.simplecloud.droplet.serverhost.runtime.config.YamlDirectoryRepository
