+++ serverhost-runtime/src/main/kotlin/app/simplecloud/droplet/serverhost/runtime/environment/MetricsTracker.kt
@@ -1,4 +1,4 @@
-package app.simplecloud.droplet.serverhost.runtime.runner
+package app.simplecloud.droplet.serverhost.runtime.environment
 
 import app.simplecloud.controller.shared.server.Server
 import app.simplecloud.droplet.api.metrics.MetricsEventNames
