+++ serverhost-runtime/src/main/kotlin/app/simplecloud/droplet/serverhost/runtime/environment/ServerEnvironment.kt
@@ -1,10 +1,8 @@
-package app.simplecloud.droplet.serverhost.runtime.runner
+package app.simplecloud.droplet.serverhost.runtime.environment
 
-import app.simplecloud.controller.shared.group.Group
 import app.simplecloud.controller.shared.server.Server
 import app.simplecloud.droplet.serverhost.runtime.config.environment.EnvironmentConfig
 import app.simplecloud.droplet.serverhost.runtime.config.environment.EnvironmentConfigRepository
-import app.simplecloud.droplet.serverhost.runtime.config.environment.BuildPolicy
 import app.simplecloud.droplet.serverhost.runtime.template.TemplateProvider
 import app.simplecloud.droplet.serverhost.shared.actions.YamlActionContext
 import app.simplecloud.droplet.serverhost.shared.actions.YamlActionTriggerTypes
@@ -25,9 +23,9 @@ abstract class ServerEnvironment(
     /**
      * Called on startup to build groups
      * @param [server] a fake server used in the build process (To make templates work correctly)
-     * @param [context] The context of the method call (can be either [BuildPolicy.ONCE] or [BuildPolicy.TRIGGER]
+     * @param [context] The context of the method call
      */
-    open fun build(server: Server, context: BuildPolicy = BuildPolicy.ONCE) {
+    open fun build(server: Server, context: BuildContext = BuildContext.STARTUP) {
 
     }
