+++ serverhost-runtime/src/main/kotlin/app/simplecloud/droplet/serverhost/runtime/environment/docker/DockerServerEnvironment.kt
@@ -1,4 +1,4 @@
-package app.simplecloud.droplet.serverhost.runtime.runner.docker
+package app.simplecloud.droplet.serverhost.runtime.environment.docker
 
 import app.simplecloud.controller.shared.host.ServerHost
 import app.simplecloud.controller.shared.server.Server
@@ -6,10 +6,7 @@ import app.simplecloud.droplet.api.time.ProtobufTimestamp
 import app.simplecloud.droplet.serverhost.runtime.config.environment.*
 import app.simplecloud.droplet.serverhost.runtime.config.environment.generators.DefaultDockerConfigGenerator
 import app.simplecloud.droplet.serverhost.runtime.launcher.ServerHostStartCommand
-import app.simplecloud.droplet.serverhost.runtime.runner.EnvironmentBuilder
-import app.simplecloud.droplet.serverhost.runtime.runner.GroupRuntimeDirectory
-import app.simplecloud.droplet.serverhost.runtime.runner.MetricsTracker
-import app.simplecloud.droplet.serverhost.runtime.runner.ServerEnvironment
+import app.simplecloud.droplet.serverhost.runtime.environment.*
 import app.simplecloud.droplet.serverhost.runtime.template.TemplateProvider
 import app.simplecloud.droplet.serverhost.shared.actions.YamlActionTriggerTypes
 import app.simplecloud.droplet.serverhost.shared.docker.DockerClientInstance
@@ -72,10 +69,10 @@ class DockerServerEnvironment(
     }
 
     @OptIn(ExperimentalPathApi::class)
-    override fun build(server: Server, context: BuildPolicy) {
+    override fun build(server: Server, context: BuildContext) {
         val env = getEnvironment(server) ?: DefaultDockerConfigGenerator.generate(args)
-        if (context == BuildPolicy.ONCE && !env.buildPolicy.firstBuild) return
-        if (context == BuildPolicy.TRIGGER && !env.buildPolicy.trigger) return
+        if (context == BuildContext.STARTUP != env.buildPolicy.firstBuild) return
+        if (context == BuildContext.TRIGGER != env.buildPolicy.trigger) return
         if (!server.properties.containsKey("docker-image")) {
             logger.error("Can not build a new docker image if none is specified")
             return
